requirements.txt
openai>=1.6.0
rdkit-pypi>=2022.9.5
matplotlib>=3.7
